fix adaptive naming
